Make peers tied to an account
Make a peer entry tied to an account. On a multi user system this is
important because a user could update another users certs and defeat
the crypto that way. With this change, peers are seperated by users. A
wildcard mode is introduced, which lets the user search across all
accounts, this is usefull for when 1 person owns all accounts in the
database.